requirements.txt
# PDF解析
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# AI调用
openai>=1.0.0
dashscope>=1.14.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0

# 数据库
SQLAlchemy>=2.0.0

# Web框架
streamlit>=1.30.0
flask>=3.0.0
flask-cors>=4.0.0

# 异步处理
aiofiles>=23.0.0
aiohttp>=3.9.0

# 工具
python-dotenv>=1.0.0

# 用户认证
bcrypt>=4.0.0

# 配置管理
pyyaml>=6.0.0

# 可视化图表（可选）
plotly>=5.18.0

# 桌面窗口启动器（选项 C）
pywebview>=4.0

